add explicit test in travis.
we need to set this automatically in cmake. But first we need a general way to exit from an application from the parallel GUI testing function. Now I use fm.close(), and I need to pass a pointer to the main form to that function to do that. A general API::exit() will be much better but don't work. Why?